refactor code, add/delete comments
